HTML Source :)
VIDEO LINK AND 6000 FONTS LINK: Video: unavaible :( (IF i can compress the video and publish it. Video is located on the files folder. So please make sure the correct folders otherwise it does NOT work.) Fonts: https://www.freepsdking.com/english-stylish-fonts-free-download-zip-file/ PASTE THE FONTS ON THE create css fonts folder and paste inside of it i got css :)